applications/onnx/requirements.txt
onnx
onnxruntime-gpu
torch
basicsr
opencv-python
numpy
pyyaml